pip_requirements.txt
# pip install torch==2.4.1+cu121 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
absl-py==2.1.0
accelerate==1.4.0
antlr4-python3-runtime==4.9.3
args==0.1.0
torch==2.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
bert-score==0.3.13
braceexpand==0.1.7
cachetools==5.5.0
charset-normalizer==3.4.1
clarabel==0.9.0
click==8.1.7
clint==0.5.1
clip==0.2.0
contourpy==1.3.0
cvxpy==1.5.3
cycler==0.12.1
datasets==3.2.0
decorator==4.4.2
decord==0.6.0
# deepseek-sdk==0.1.0
# deepseek-vl==0.0.1.dev0
diffusers==0.30.3
diskcache==5.6.3
docker-pycreds==0.4.0
docstring_parser==0.16
ecos==2.0.14
einops==0.8.0
et_xmlfile==2.0.0
fairscale==0.4.13
filelock==3.17.0
fire==0.4.0
# flash-attn==2.7.2.post1
flyingsquid==0.0.0a0
fonttools==4.54.1
fsspec
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
hpack==4.0.0
# hpsv2==1.2.0
httplib2==0.22.0
huggingface-hub==0.29.2
hyperlm==0.0.5
# image-reward==1.5
imageio==2.36.0
imageio-ffmpeg==0.5.1
inflect==7.5.0
iniconfig==2.0.0
iopath==0.1.10
jaxtyping==0.2.36
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.11.9
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.9.2
maturin==1.8.2
mdurl==0.1.2
ml_collections==1.0.0
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
munkres==1.1.4
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numbskull==0.1.1
numpy==2.0.0
omegaconf==2.3.0
# open_clip_torch==2.30.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
osqp==0.6.7.post1
patsy==0.5.6
peft==0.13.2
pgmpy==0.1.26
plotly==5.23.0
pluggy==1.5.0
portalocker==2.10.1
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.29.3
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.8
pyparsing==3.1.4
pytest==7.2.0
pytest-split==0.8.0
pytokenizations==0.8.4
pytorch-lightning==2.5.0.post0
PyYAML==6.0.2
qdldl==0.1.7.post4
regex==2024.11.6
requests==2.32.3
rich==13.9.2
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scs==3.2.7
seaborn==0.13.2
sentence-transformers==3.4.1
sentry-sdk==2.17.0
setproctitle==1.3.3
shtab==1.7.1
smmap==5.0.1
snorkel==0.9.9
snorkel-metal==0.5.0
soupsieve==2.6
statsmodels==0.14.4
sty==1.0.6
sympy==1.13.1
# t2v_metrics==1.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.15
tokenizers==0.21.0
tomli==2.2.1
torch-fidelity==0.3.0
torchmetrics==1.6.1
torchvision
transformers==4.48.3
transformers-stream-generator==0.0.5
trl==0.11.4
typeguard==4.4.2
tyro==0.8.12
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
wandb==0.18.3
webdataset==0.2.100
Werkzeug==3.0.4
xgboost==2.1.1